+++ setupext/test_java.py
@@ -29,6 +29,7 @@ def compileJava():
         pass
     version = getJavaVersion(javac)
     srcs = glob.glob('test/harness/jpype/**/*.java', recursive=True)
+    srcs.extend(glob.glob('test/harness/org/**/*.java', recursive=True))
     exports = ""
     if version > 7:
         srcs.extend(glob.glob('test/harness/java8/**/*.java', recursive=True))